AG-11304 - Move doc interface patching to the generation step
