+++ packages/ag-charts-website/src/features/api-documentation/apiReferenceHelpers.ts
@@ -1,5 +1,4 @@
 import type { ApiReferenceConfig } from '@features/api-documentation/components/ApiReference';
-
 import type {
     ApiReferenceNode,
     ApiReferenceType,
@@ -8,7 +7,7 @@ import type {
     MemberNode,
     TypeLiteralNode,
     TypeNode,
-} from './api-reference-types';
+} from '@generate-code-reference-plugin/doc-interfaces/types';
 
 type PossibleTypeNode = TypeNode | undefined | PossibleTypeNode[];
 
@@ -297,98 +296,7 @@ export function extractSearchData(
     return [];
 }
 
-export function patchAgChartOptionsReference(reference: ApiReferenceType) {
-    const interfaceRef = reference.get('AgChartOptions');
-    const getTypeUnion = (typeRef: ApiReferenceNode | undefined): string[] => {
-        if (typeRef?.kind === 'typeAlias') {
-            if (typeof typeRef.type === 'string') {
-                return [typeRef.type];
-            }
-            if (typeof typeRef.type === 'object' && typeRef.type.kind === 'union') {
-                return typeRef.type.type.filter((type): type is string => typeof type === 'string');
-            }
-        }
-        return [];
-    };
-
-    const axisOptions: string[] = [];
-    const seriesOptions: string[] = [];
-
-    let altInterface: InterfaceNode | null = null;
-
-    for (const typeName of getTypeUnion(interfaceRef)) {
-        const typeRef = reference.get(typeName);
-
-        if (typeRef?.kind !== 'interface') {
-            throw Error('Unexpected AgChartOptions union type');
-        }
-
-        altInterface ??= typeRef;
-
-        for (const member of typeRef.members) {
-            if (
-                typeof member.type !== 'object' ||
-                member.type.kind !== 'array' ||
-                typeof member.type.type !== 'string'
-            ) {
-                continue;
-            }
-
-            if (member.name === 'axes') {
-                const axisOption = reference.get(member.type.type);
-                // Ignore `groupedCategory` axis, as it is an implementation detail
-                const axisOptionUnion = getTypeUnion(axisOption).filter(
-                    (name) => name !== 'AgGroupedCategoryAxisOptions'
-                );
-                axisOptions.push(...axisOptionUnion);
-            } else if (member.name === 'series') {
-                seriesOptions.push(...getTypeUnion(reference.get(member.type.type)));
-            }
-        }
-    }
-
-    if (altInterface === null) {
-        return;
-    }
-
-    reference.set('AgChartAxisOptions', {
-        kind: 'typeAlias',
-        name: 'AgChartAxisOptions',
-        type: { kind: 'union', type: axisOptions },
-    });
-    reference.set('AgChartSeriesOptions', {
-        kind: 'typeAlias',
-        name: 'AgChartSeriesOptions',
-        type: { kind: 'union', type: seriesOptions },
-    });
-
-    altInterface = {
-        ...altInterface,
-        name: 'AgChartOptions',
-        members: altInterface.members.map((member) => {
-            if (typeof member.type !== 'object') {
-                return member;
-            }
-            if (member.name === 'axes') {
-                return Object.assign({}, member, {
-                    type: Object.assign({}, member.type, { type: 'AgChartAxisOptions' }),
-                });
-            }
-            if (member.name === 'series') {
-                return Object.assign({}, member, {
-                    type: Object.assign({}, member.type, { type: 'AgChartSeriesOptions' }),
-                });
-            }
-            return member;
-        }),
-    };
-
-    reference.set('AgChartOptions', altInterface);
-}
-
 export function getOptionsStaticPaths(reference: ApiReferenceType) {
-    patchAgChartOptionsReference(reference);
-
     const getSubTypes = (ref: ApiReferenceNode): string[] =>
         ref.kind === 'typeAlias' &&
         typeof ref.type === 'object' &&
