+++ packages/ag-charts-website/src/pages/options.astro
@@ -1,11 +1,9 @@
 ---
-import { patchAgChartOptionsReference } from '@features/api-documentation/apiReferenceHelpers';
 import { ApiReferencePage } from '@features/api-documentation/components/ApiReferencePage';
 import APIViewLayout from '@layouts/APIViewLayout.astro';
 import { getInterfacesReference } from '@utils/server/getInterfacesReference';
 
 const reference = getInterfacesReference();
-patchAgChartOptionsReference(reference);
 ---
 
 <APIViewLayout title="Options API">
