+++ packages/ag-charts-website/src/pages/options/[memberName]/[type].astro
@@ -1,5 +1,5 @@
 ---
-import { getOptionsStaticPaths, patchAgChartOptionsReference } from '@features/api-documentation/apiReferenceHelpers';
+import { getOptionsStaticPaths } from '@features/api-documentation/apiReferenceHelpers';
 import { ApiReferencePage } from '@features/api-documentation/components/ApiReferencePage';
 import APIViewLayout from '@layouts/APIViewLayout.astro';
 import { getInterfacesReference } from '@utils/server/getInterfacesReference';
@@ -10,7 +10,6 @@ export async function getStaticPaths() {
 }
 
 const reference = getInterfacesReference();
-patchAgChartOptionsReference(reference);
 ---
 
 <APIViewLayout title="Options API">
