--- packages/ag-charts-website/src/utils/hidden-api-options.json
@@ -1,12 +1,0 @@
-{
-    "AgGradientLegendOptions": ["stop"],
-    "AgHeatmapSeriesOptions": ["showInLegend"],
-    "AgSunburstSeriesHighlightStyle": ["item", "series"],
-    "AgSunburstSeriesOptions": ["showInLegend"],
-    "AgTreemapSeriesHighlightStyle": ["item", "series"],
-    "AgTreemapSeriesOptions": ["showInLegend"],
-    "AgAngleCategoryAxisOptions": ["keys"],
-    "AgAngleNumberAxisOptions": ["keys"],
-    "AgRadiusCategoryAxisOptions": ["keys"],
-    "AgRadiusNumberAxisOptions": ["keys"]
-}
