+++ packages/ag-charts-website/src/utils/pages.ts
@@ -34,7 +34,7 @@ export interface DevFileRoute {
  * NOTE: File path is after `getRootUrl()`
  */
 export const DEV_FILE_PATH_MAP: Record<string, string> = {
-    'resolved-interfaces.json': 'dist/packages/ag-charts-community/resolved-interfaces.AUTO.json',
+    'resolved-interfaces-patched.json': 'dist/packages/ag-charts-community/resolved-interfaces-patched.AUTO.json',
 
     'ag-charts-community/dist/**': 'packages/ag-charts-community/dist/**/*.{cjs,js,map}',
     'ag-charts-enterprise/dist/**': 'packages/ag-charts-enterprise/dist/**/*.{cjs,js,map}',
