+++ packages/ag-charts-website/src/utils/server/getInterfacesReference.ts
@@ -1,7 +1,6 @@
-import type { ApiReferenceType } from '@features/api-documentation/api-reference-types';
+import type { ApiReferenceType } from '@generate-code-reference-plugin/doc-interfaces/types';
 import { DEV_FILE_PATH_MAP, getRootUrl } from '@utils/pages';
 import { pathJoin } from '@utils/pathJoin';
-import { updateInterfaceReferences } from '@utils/updateInterfaceReferences';
 import { readFileSync } from 'node:fs';
 
 type DevFileKey = keyof typeof DEV_FILE_PATH_MAP;
@@ -18,6 +17,6 @@ function getJsonFromDevFile(devFileKey: DevFileKey) {
  * For static site generation
  */
 export function getInterfacesReference(): ApiReferenceType {
-    const jsonContent = getJsonFromDevFile('resolved-interfaces.json');
-    return updateInterfaceReferences(jsonContent);
+    const jsonContent = getJsonFromDevFile('resolved-interfaces-patched.json');
+    return new Map(Object.entries(jsonContent));
 }
