+++ plugins/ag-charts-generate-code-reference-files/src/doc-interfaces/constants.ts
@@ -1,0 +1,12 @@
+export const HIDDEN_API_OPTIONS = {
+    AgGradientLegendOptions: ['stop'],
+    AgHeatmapSeriesOptions: ['showInLegend'],
+    AgSunburstSeriesHighlightStyle: ['item', 'series'],
+    AgSunburstSeriesOptions: ['showInLegend'],
+    AgTreemapSeriesHighlightStyle: ['item', 'series'],
+    AgTreemapSeriesOptions: ['showInLegend'],
+    AgAngleCategoryAxisOptions: ['keys'],
+    AgAngleNumberAxisOptions: ['keys'],
+    AgRadiusCategoryAxisOptions: ['keys'],
+    AgRadiusNumberAxisOptions: ['keys'],
+};
